WIP - Ensure that the image variants are created for the PortletRepository too (it's used by the blogs portlet to store the images)
