+++ adaptive-media-document-library/bnd.bnd
@@ -3,3 +3,4 @@ Bundle-SymbolicName: com.liferay.adaptive.media.document.library
 Bundle-Version: 1.0.0
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Adaptive Media
+-dsannotations-options: inherit
