+++ adaptive-media-document-library/src/main/java/com/liferay/adaptive/media/document/library/repository/internal/AdaptiveMediaLiferayRepositoryDefiner.java
@@ -1,0 +1,43 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.adaptive.media.document.library.repository.internal;
+
+import com.liferay.portal.kernel.repository.registry.RepositoryDefiner;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+
+/**
+ * @author Adolfo Pérez
+ */
+@Component(immediate = true, service = RepositoryDefiner.class)
+public class AdaptiveMediaLiferayRepositoryDefiner
+	extends BaseOverridingRepositoryDefiner {
+
+	@Activate
+	protected void activate() {
+		initializeOverridenRepositoryDefiner(_CLASS_NAME);
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		restoreOverridenRepositoryDefiner(_CLASS_NAME);
+	}
+
+	private static final String _CLASS_NAME =
+		"com.liferay.portal.repository.liferayrepository.LiferayRepository";
+
+}
